#0 Fixed a bug in how the default proxy is formed.
The way that Globus (grid-proxy-init and grid-proxy-info at least) determine the
proxy file is by using getuid().  The $USER environment variable is not
protected, and it can be set to any value, or not set at all.  Instead of
getting the UID by calling getpwnam($USER), it should rely on either $< or $>
(real or effective user ID, whichever one Globus uses).

For full details see:
https://ticket.grid.iu.edu/goc/viewer?id=9034


git-svn-id: https://vdt.cs.wisc.edu/svn/rsv/branches/rsv-test@11171 4e558342-562e-0410-864c-e07659590f8c