#0 Remove the osg-directories-diskspace-probe.  It's not useful:  - it doesn't actually give any warnings/errors  - it doesn't work with NFS filesystems because it's not flexible enough when    parsing 'df' output
I think it was basically a beta-quality probe that never got completed.  It was
disabled by default so probably nobody has ever run it.

Also removed the worker script it used which implemented both this probe and the
directories-permissions probe (which we have rewritten and no longer needs the
worker script)


git-svn-id: https://vdt.cs.wisc.edu/svn/rsv/branches/rsv-test@11198 4e558342-562e-0410-864c-e07659590f8c