#0 Merged Anand's change r11834 from rsv-security to rsv-test (with one tweak):
His commit deleted the md5sum-ca-verify.sh script but I preserved it since there
are still references to the script in the code.  He said that these references
are no longer used but I didn't want to take the time to validate and it does
not hurt to leave the script in place for now.  We'll get rid of it if/when we
transition to "new" style RSV scripts that don't rely on RSVProbeBase.pm


git-svn-id: https://vdt.cs.wisc.edu/svn/rsv/branches/rsv-test@11893 4e558342-562e-0410-864c-e07659590f8c