Use 'gatekeeper-type' instead of 'ce-type' or 'condor-ce'
Rename --ce-type to --gatekeeper-type in rsv-control

The latter name is more accurate. Also, the values it takes are 'gram' and
'condor-ce' instead of 'gram' and 'htcondor', for consistency.

Replace 'condor-ce' metric option replaced with 'gatekeeper-type'

The boolean option 'condor-ce' for each metric has been replaced with the
'gatekeeper-type' option, which takes values 'gram' and 'condor-ce'. This is
for consistency with the command-line argument to rsv-control.
(SOFTWARE-1048)

git-svn-id: https://vdt.cs.wisc.edu/svn/software/rsv/trunk@17455 4e558342-562e-0410-864c-e07659590f8c