fixup for #9 (SOFTWARE-2782)
There were two things wrong with the first approach:

The "success" condition was supposed to be when ccq_output was empty,
but in that case `echo | wc -l` still returned 1, since echo outputs
an empty line.  Piping the condor_cron_q output directly into wc -l
does the right thing here.

But also, ccq_output gets used in the output messages for diagnostic
purposes, and a list of 1's is not as useful as the default output.